clean code and fix CRUD in 2022-11-10

+++ src/main/java/com/vti/testing/service/UserService.java
@@ -1,16 +1,23 @@
 package com.vti.testing.service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import com.vti.testing.dto.UserDTO;
+import com.vti.testing.entity.Role;
 import com.vti.testing.entity.User;
 import com.vti.testing.exception.custom_exception.NotFoundEx;
+import com.vti.testing.formcreate.FormUserCreate;
+import com.vti.testing.repository.IRoleRepository;
 import com.vti.testing.repository.IUserRepository;
 
+import com.vti.testing.responseobj.ResponseObj;
+import com.vti.testing.service.interfaces.IUserService;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 @Service
 public class UserService implements IUserService {
@@ -18,7 +25,11 @@ public class UserService implements IUserService {
 	@Autowired
 	private IUserRepository userRepository;
 	@Autowired
+	private IRoleRepository roleRepository;
+	@Autowired
 	private ModelMapper modelMapper;
+	@Autowired
+	private BCryptPasswordEncoder passwordEncoder;
 	
 	@Override
 	public User getByUserName(String userName) {
@@ -35,17 +46,27 @@ public class UserService implements IUserService {
 		throw new NotFoundEx("Cannot find user with this id " + id );
 	}
 
-//	@Override
-//	public void newUser(FormUserCreate newUser) {
-//		User user = new User(newUser.getUserName(), newUser.getPassWord(), newUser.getActive(), newUser.getRole());
-//		userRepository.save(user);
-//
-//	}
+	@Override
+	public ResponseObj newUser(FormUserCreate newUser) {
+		// Encode passWord
+		String passEncode = passwordEncoder.encode(newUser.getPassWord());
+		newUser.setPassWord(passEncode);
+
+		User user = new User(newUser.getUserName(), newUser.getPassWord());
+		Role role = roleRepository.findByRoleName("ROLE_USER");
+		List<Role> newUserRoles = new ArrayList<Role>();
+			newUserRoles.add(role); // role default = ROLE_USER
+		user.setRoles(newUserRoles);
+		user.setActive("1");
+		userRepository.save(user);
+		return new ResponseObj("200", "Created New User", user.getUserName());
+	}
 
 	@Override
 	public List<UserDTO> getAllUsers() {
-		List<User> entitys = userRepository.findAll();
-		List<UserDTO> dtos = modelMapper.map(entitys,
+		List<User> entities = userRepository.findAll();
+
+		List<UserDTO> dtos = modelMapper.map(entities,
 				new TypeToken<List<UserDTO>>(){}.getType());
 		return dtos;
 	}
@@ -59,8 +80,14 @@ public class UserService implements IUserService {
 	}
 
 	@Override
-	public void deleteUser(int id) {
+	public ResponseObj deleteUser(int id) {
+		if (userRepository.existsById(id)){
+			String userNameDelete = userRepository.findById(id).get().getUserName();
 			userRepository.deleteById(id);
+			return new ResponseObj("200", "Delete Success !", userNameDelete);
+		}
+		throw new NotFoundEx("User is not exists with this id");
+
 	}
 
 }
